% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{blockr.sdtm-package}
\alias{blockr.sdtm}
\alias{blockr.sdtm-package}
\title{blockr.sdtm: SDTM data for blockr}
\description{
Use SDTM data as provided by 'pharmaversesdtm' as part of a blockr pipeline.
}
\author{
\strong{Maintainer}: Nicolas Bennett \email{nicolas@cynkra.com}

}
\keyword{internal}
