% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdtm.R
\name{new_sdtm_block}
\alias{new_sdtm_block}
\title{SDTM data block constructor}
\usage{
new_sdtm_block(dataset = character(), ...)
}
\arguments{
\item{dataset}{Selected dataset}

\item{...}{Forwarded to \code{\link[=new_block]{new_block()}}}
}
\description{
This block allows to selected data from a package,
the default being datasets.
}
